requirements.txt
flask
requests
joblib
pandas
scikit-learn
numpy
openpyxl
python-dotenv